Renames and added helper method for block creation
